refactor: move PropertyType to its own folder
