+++ packages/f36-docs-utils/src/Props/PropertyType/index.ts
@@ -1,0 +1,1 @@
+export { PropertyType } from './PropertyType';
